requirements.txt
pyspark>=3.4.0
pandas>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
pycountry>=23.12.0
iso3166>=2.1.0
